# artifacts/api-server/src/routes/recipes.ts
import { openai } from "@workspace/integrations-openai-ai-server";
import { db, recipeNutritionTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { Router, type IRouter } from "express";
import { logger } from "../lib/logger";

const router: IRouter = Router();
const API_ROOT = "https://www.themealdb.com/api/json/v1/1";
const SOURCE = "TheMealDB";
const SOURCE_URL = "https://www.themealdb.com/";

// ─── Nutrition estimation ────────────────────────────────────────────────────

type NutritionEstimate = { calories: number; proteinG: number; carbsG: number; fatG: number };

// Maximum time to wait for an OpenAI response before giving up.  8 s is
// generous enough for a well-behaved call while still preventing the HTTP
// request from hanging indefinitely when OpenAI is slow or unreachable.
// Tests may shorten this via OPENAI_TIMEOUT_MS_OVERRIDE so they run without
// fake timers and still validate the abort/fallback path quickly.
const OPENAI_TIMEOUT_MS =
  process.env.OPENAI_TIMEOUT_MS_OVERRIDE
    ? Number(process.env.OPENAI_TIMEOUT_MS_OVERRIDE)
    : 8_000;

// TTL for database-persisted nutrition estimates.  7 days is long enough that
// popular meals are rarely re-estimated, but short enough that upstream
// ingredient changes from TheMealDB surface within a reasonable window.
const NUTRITION_DB_TTL_MS = 1000 * 60 * 60 * 24 * 7; // 7 days

// L1: in-memory cache with a timestamp so staleness can be checked without a
// DB round-trip.  The timestamp reflects when the estimate was originally
// computed (i.e. we carry the DB row age into L1 so a stale row is also stale
// in memory after the server restarts and reads the row back).
type CachedEntry = { estimate: NutritionEstimate; cachedAt: number };
const nutritionCache = new Map<string, CachedEntry>();

// Single-flight guard: prevents two concurrent background refreshes for the
// same meal from both hitting OpenAI simultaneously.
const nutritionRefreshInFlight = new Set<string>();

/** Look up a persisted estimate from the database (L2 cache), with staleness info.
 *  Returns the original `createdAtMs` epoch so callers can propagate it into
 *  the L1 cache — this ensures the L1 TTL is anchored to when the estimate was
 *  computed, not to when the DB row was read. */
async function getNutritionFromDb(
  mealId: string,
): Promise<{ estimate: NutritionEstimate; createdAtMs: number; isStale: boolean } | null> {
  try {
    const rows = await db
      .select()
      .from(recipeNutritionTable)
      .where(eq(recipeNutritionTable.mealId, mealId))
      .limit(1);
    const row = rows[0];
    if (!row) return null;
    const createdAtMs = row.createdAt.getTime();
    const age = Date.now() - createdAtMs;
    return {
      estimate: { calories: row.calories, proteinG: row.proteinG, carbsG: row.carbsG, fatG: row.fatG },
      createdAtMs,
      isStale: age > NUTRITION_DB_TTL_MS,
    };
  } catch (err) {
    logger.warn({ err, mealId }, "nutrition DB read failed — falling back to OpenAI");
    return null;
  }
}

/** Persist an estimate to the database, overwriting any existing row so that
 *  refreshed estimates always replace the stale one rather than being dropped. */
async function saveNutritionToDb(mealId: string, nutrition: NutritionEstimate): Promise<void> {
  try {
    await db
      .insert(recipeNutritionTable)
      .values({ mealId, ...nutrition })
      .onConflictDoUpdate({
        target: recipeNutritionTable.mealId,
        set: { ...nutrition, createdAt: new Date() },
      });
  } catch (err) {
    // Best-effort — a write failure should never break the response.
    logger.warn({ err, mealId }, "nutrition DB write failed — estimate not persisted");
  }
}

async function estimateNutrition(name: string, ingredients: string[]): Promise<NutritionEstimate | null> {
  const controller = new AbortController();
  const timer = setTimeout(() => controller.abort(), OPENAI_TIMEOUT_MS);
  try {
    const completion = await openai.chat.completions.create(
      {
        model: "gpt-5.4-mini",
        messages: [
          {
            role: "system",
            content:
              "You are a nutrition expert. Return ONLY a JSON object — no markdown, no prose — with these four integer keys: calories, proteinG, carbsG, fatG. Estimate values for one typical serving.",
          },
          {
            role: "user",
            content: `Recipe: ${name}\nIngredients: ${ingredients.join(", ")}`,
          },
        ],
        response_format: { type: "json_object" },
        max_tokens: 80,
      },
      { signal: controller.signal },
    );
    const raw = completion.choices[0]?.message?.content ?? "{}";
    const parsed = JSON.parse(raw) as Record<string, unknown>;
    const calories = Math.round(Number(parsed.calories) || 0);
    const proteinG = Math.round(Number(parsed.proteinG) || 0);
    const carbsG = Math.round(Number(parsed.carbsG) || 0);
    const fatG = Math.round(Number(parsed.fatG) || 0);
    if (calories <= 0) return null; // nonsensical estimate — skip
    return { calories, proteinG, carbsG, fatG };
  } catch {
    return null;
  } finally {
    clearTimeout(timer);
  }
}

/**
 * Re-estimate nutrition for a meal in the background and update both caches.
 * Clears the in-flight guard when done (success or failure).
 */
async function refreshNutritionInBackground(
  mealId: string,
  name: string,
  ingredients: string[],
): Promise<void> {
  try {
    const fresh = await estimateNutrition(name, ingredients);
    if (fresh) {
      nutritionCache.set(mealId, { estimate: fresh, cachedAt: Date.now() });
      void saveNutritionToDb(mealId, fresh);
      logger.info({ mealId }, "nutrition estimate refreshed in background");
    }
  } catch (err) {
    logger.warn({ err, mealId }, "background nutrition refresh failed");
  } finally {
    nutritionRefreshInFlight.delete(mealId);
  }
}

/**
 * Resolve nutrition for a meal, checking caches in order:
 *   L1 (in-memory Map) → L2 (database) → OpenAI estimation
 *
 * Staleness policy:
 *   - A fresh hit (age < NUTRITION_DB_TTL_MS) is returned immediately.
 *   - A stale hit is returned immediately so the card shows data at once, but
 *     a background re-estimation is triggered so the next request (or the
 *     next server start) gets a fresh value.  The single-flight guard
 *     (nutritionRefreshInFlight) ensures only one refresh per meal runs at
 *     a time even if many concurrent requests hit the same stale entry.
 */
async function resolveNutrition(
  mealId: string,
  name: string,
  ingredients: string[],
): Promise<NutritionEstimate | null> {
  // L1 hit — check freshness
  const memHit = nutritionCache.get(mealId);
  if (memHit) {
    const isStale = Date.now() - memHit.cachedAt > NUTRITION_DB_TTL_MS;
    if (isStale && !nutritionRefreshInFlight.has(mealId) && ingredients.length > 0) {
      nutritionRefreshInFlight.add(mealId);
      void refreshNutritionInBackground(mealId, name, ingredients);
    }
    // Always return the current value — fresh or stale — so the UI is never blank.
    return memHit.estimate;
  }

  // L2 hit — carry the row's original createdAt into L1 so the L1 TTL is
  // anchored to when the estimate was computed, not when the row was read.
  // A row that is 6.9 days old therefore expires in L1 after only ~2.4 hours,
  // rather than receiving a fresh 7-day lease.
  const dbResult = await getNutritionFromDb(mealId);
  if (dbResult) {
    nutritionCache.set(mealId, { estimate: dbResult.estimate, cachedAt: dbResult.createdAtMs });

    if (dbResult.isStale && !nutritionRefreshInFlight.has(mealId) && ingredients.length > 0) {
      nutritionRefreshInFlight.add(mealId);
      void refreshNutritionInBackground(mealId, name, ingredients);
    }
    return dbResult.estimate;
  }

  // Cache miss — call OpenAI
  if (ingredients.length === 0) return null;
  const fresh = await estimateNutrition(name, ingredients);
  if (fresh) {
    nutritionCache.set(mealId, { estimate: fresh, cachedAt: Date.now() });
    void saveNutritionToDb(mealId, fresh);
  }
  return fresh;
}

// ─── Background nutrition warm-up ────────────────────────────────────────────

const WARMUP_BATCH_SIZE = 18;
const WARMUP_DELAY_MS = 500; // 500 ms between OpenAI calls to stay well within quota

// Single-flight guard: only one warm-up job may run at a time.
let warmupInProgress = false;
// Flips to true once the first warm-up job finishes; resets when the pool TTL
// expires so a fresh pool always triggers a new warm cycle.
let warmupDone = false;

/**
 * Silently pre-populate the nutrition cache for the first page of the "For you"
 * pool.  Runs entirely in the background — callers must NOT await it.
 *
 * Strategy:
 *   1. Take up to WARMUP_BATCH_SIZE meals that are not already in L1 (memory).
 *   2. For each, check L2 (DB) before hitting OpenAI.
 *   3. A full-detail fetch is required for ingredients (the pool only carries
 *      summary fields from the category filter endpoint).
 *   4. 500 ms delay between OpenAI calls prevents quota bursting.
 */
async function warmNutritionCache(meals: Meal[]): Promise<void> {
  // Single-flight: bail if a job is already running.
  if (warmupInProgress) return;
  warmupInProgress = true;

  // Prefer meals at the front of the pool; skip any already in L1.
  const candidates = meals
    .filter((m) => !nutritionCache.has(m.idMeal))
    .slice(0, WARMUP_BATCH_SIZE);

  if (candidates.length === 0) {
    warmupDone = true;
    warmupInProgress = false;
    return;
  }

  logger.info({ count: candidates.length }, "nutrition warm-up started");
  let warmed = 0;

  for (const meal of candidates) {
    try {
      // Re-check L1 — a concurrent detail request may have filled it since.
      if (nutritionCache.has(meal.idMeal)) continue;

      // L2 check — use a fresh DB hit without touching OpenAI; skip stale rows
      // (they will be refreshed lazily when a user opens the detail sheet).
      // Carry the row's original createdAt into L1 so the TTL is anchored to
      // the estimation time, not the current clock.
      const dbResult = await getNutritionFromDb(meal.idMeal);
      if (dbResult && !dbResult.isStale) {
        nutritionCache.set(meal.idMeal, { estimate: dbResult.estimate, cachedAt: dbResult.createdAtMs });
        continue;
      }

      // Need the full meal record to extract ingredients.
      const data = await fetchJson(`${API_ROOT}/lookup.php?i=${encodeURIComponent(meal.idMeal)}`);
      const fullMeal = data.meals?.[0];
      if (!fullMeal) continue;

      const recipe = toRecipe(fullMeal);
      if (recipe.ingredients.length === 0) continue;

      const nutrition = await estimateNutrition(recipe.name, recipe.ingredients);
      if (nutrition) {
        nutritionCache.set(meal.idMeal, { estimate: nutrition, cachedAt: Date.now() });
        void saveNutritionToDb(meal.idMeal, nutrition);
        warmed++;
      }

      // Rate-limit: pause between OpenAI requests.
      await new Promise<void>((resolve) => setTimeout(resolve, WARMUP_DELAY_MS));
    } catch (err) {
      logger.warn({ err, mealId: meal.idMeal }, "nutrition warm-up skipped meal");
    }
  }

  warmupDone = true;
  warmupInProgress = false;
  logger.info({ warmed }, "nutrition warm-up complete");
}

// ─── "For you" pool ───────────────────────────────────────────────────────────

// Categories fetched in parallel to build the "For you" pool.
// Ordered so the feed opens with variety: lighter meals first, then
// heartier ones, then dessert — a natural browse rhythm.
const FOR_YOU_CATEGORIES = [
  "Vegetarian", "Seafood", "Chicken", "Pasta",
  "Beef", "Lamb", "Miscellaneous", "Side", "Dessert", "Starter",
];
const FOR_YOU_TTL_MS = 1000 * 60 * 60; // 1 hour

let forYouCache: Meal[] = [];
let forYouCacheTime = 0;
let forYouFetchPromise: Promise<Meal[]> | null = null;

type Meal = {
  idMeal: string;
  strMeal: string;
  strMealThumb?: string | null;
  strCategory?: string | null;
  strArea?: string | null;
  strInstructions?: string | null;
  strTags?: string | null;
  [key: string]: string | null | undefined;
};

function toRecipe(meal: Meal) {
  const ingredients = Array.from({ length: 20 }, (_, index) => {
    const ingredient = meal[`strIngredient${index + 1}`]?.trim();
    const measure = meal[`strMeasure${index + 1}`]?.trim();
    return ingredient ? `${measure ? `${measure} ` : ""}${ingredient}` : null;
  }).filter((value): value is string => Boolean(value));

  return {
    id: meal.idMeal,
    name: meal.strMeal,
    image: meal.strMealThumb ?? null,
    category: meal.strCategory ?? null,
    area: meal.strArea ?? null,
    description: null,
    instructions: meal.strInstructions ?? null,
    ingredients,
    tags: meal.strTags ? meal.strTags.split(",").map((tag) => tag.trim()).filter(Boolean) : [],
    prepMinutes: null,
    calories: null,
    proteinG: null,
    carbsG: null,
    fatG: null,
    source: SOURCE,
    sourceUrl: SOURCE_URL,
  };
}

async function fetchJson(url: string) {
  const response = await fetch(url);
  if (!response.ok) throw new Error(`Recipe provider returned ${response.status}`);
  return response.json() as Promise<{ meals?: Meal[] | null }>;
}

async function buildForYouPool(): Promise<Meal[]> {
  const results = await Promise.allSettled(
    FOR_YOU_CATEGORIES.map((cat) => fetchJson(`${API_ROOT}/filter.php?c=${encodeURIComponent(cat)}`)),
  );
  // Interleave results so the feed has variety on every page rather than
  // showing one whole category then the next.
  const buckets = results.map((r) => (r.status === "fulfilled" ? (r.value.meals ?? []) : []));
  const maxLen = Math.max(...buckets.map((b) => b.length), 0);
  const interleaved: Meal[] = [];
  for (let i = 0; i < maxLen; i++) {
    for (const bucket of buckets) {
      if (i < bucket.length) interleaved.push(bucket[i]!);
    }
  }
  // Deduplicate by idMeal (same dish can appear in multiple categories)
  const seen = new Set<string>();
  return interleaved.filter((m) => {
    if (seen.has(m.idMeal)) return false;
    seen.add(m.idMeal);
    return true;
  });
}

async function getForYouMeals(): Promise<Meal[]> {
  const now = Date.now();
  if (forYouCache.length > 0 && now - forYouCacheTime < FOR_YOU_TTL_MS) return forYouCache;
  // Coalesce concurrent requests into one fetch
  if (!forYouFetchPromise) {
    forYouFetchPromise = buildForYouPool().then((meals) => {
      forYouCache = meals;
      forYouCacheTime = Date.now();
      forYouFetchPromise = null;
      // Reset warm-up state so the fresh pool always triggers a new cycle.
      warmupDone = false;
      warmupInProgress = false;
      // Fire-and-forget: warm the nutrition cache so first-visit cards show
      // calorie estimates without the user ever opening a detail sheet.
      void warmNutritionCache(meals);
      return meals;
    }).catch((err) => {
      forYouFetchPromise = null;
      throw err;
    });
  }
  return forYouFetchPromise;
}

router.get("/v1/recipes", async (req, res) => {
  try {
    const query = typeof req.query.query === "string" ? req.query.query.trim() : "";
    const category = typeof req.query.category === "string" ? req.query.category.trim() : "";
    const parsedLimit = Number(req.query.limit ?? 12);
    const limit = Number.isFinite(parsedLimit) ? Math.min(Math.max(parsedLimit, 1), 30) : 12;
    const parsedOffset = Number(req.query.offset ?? 0);
    const offset = Number.isFinite(parsedOffset) ? Math.max(Math.floor(parsedOffset), 0) : 0;

    let meals: Meal[];
    if (query) {
      const data = await fetchJson(`${API_ROOT}/search.php?s=${encodeURIComponent(query)}`);
      meals = data.meals ?? [];
    } else if (category) {
      const data = await fetchJson(`${API_ROOT}/filter.php?c=${encodeURIComponent(category)}`);
      meals = data.meals ?? [];
    } else {
      // "For you" — multi-category interleaved pool, cached server-side
      meals = await getForYouMeals();
    }

    const recipes = meals.slice(offset, offset + limit).map((meal) => {
      const recipe = toRecipe(meal);
      // Attach any L1-cached estimate so the card can show ~kcal without
      // the user needing to open the detail sheet first.
      const cached = nutritionCache.get(recipe.id);
      return cached ? { ...recipe, ...cached.estimate } : recipe;
    });
    // Let clients know they should refetch soon if the background warm-up
    // has not yet populated estimates for the first page of results.
    const warmupPending = !warmupDone;
    res.json({ source: SOURCE, recipes, warmupPending });
  } catch (error) {
    res.status(502).json({ message: error instanceof Error ? error.message : "Recipe provider unavailable" });
  }
});

router.get("/v1/recipes/:recipeId", async (req, res) => {
  try {
    const data = await fetchJson(`${API_ROOT}/lookup.php?i=${encodeURIComponent(req.params.recipeId)}`);
    const meal = data.meals?.[0];
    if (!meal) {
      res.status(404).json({ message: "Recipe not found" });
      return;
    }
    const base = toRecipe(meal);

    // ── Fast path: check L1 (memory) — no I/O needed ─────────────────────
    const memHit = nutritionCache.get(base.id);
    if (memHit) {
      const isStale = Date.now() - memHit.cachedAt > NUTRITION_DB_TTL_MS;
      if (isStale && !nutritionRefreshInFlight.has(base.id) && base.ingredients.length > 0) {
        nutritionRefreshInFlight.add(base.id);
        void refreshNutritionInBackground(base.id, base.name, base.ingredients);
      }
      res.json({ ...base, ...memHit.estimate });
      return;
    }

    // ── L2: check DB — fast (~10 ms), no AI call ──────────────────────────
    const dbResult = await getNutritionFromDb(base.id);
    if (dbResult) {
      nutritionCache.set(base.id, { estimate: dbResult.estimate, cachedAt: dbResult.createdAtMs });
      if (dbResult.isStale && !nutritionRefreshInFlight.has(base.id) && base.ingredients.length > 0) {
        nutritionRefreshInFlight.add(base.id);
        void refreshNutritionInBackground(base.id, base.name, base.ingredients);
      }
      res.json({ ...base, ...dbResult.estimate });
      return;
    }

    // ── Cache miss: call OpenAI now so this response includes nutrition ─────
    // We await the estimate here rather than deferring it — callers should
    // always get a value on first fetch when OpenAI is reachable.  The result
    // is written to both L1 and L2 so subsequent requests (and server restarts)
    // are served from cache without a further OpenAI call.
    if (base.ingredients.length > 0) {
      const fresh = await estimateNutrition(base.name, base.ingredients);
      if (fresh) {
        nutritionCache.set(base.id, { estimate: fresh, cachedAt: Date.now() });
        void saveNutritionToDb(base.id, fresh);
        res.json({ ...base, ...fresh });
        return;
      }
    }
    res.json({ ...base, nutritionPending: true });
    return;
  } catch (error) {
    res.status(502).json({ message: error instanceof Error ? error.message : "Recipe provider unavailable" });
    return;
  }
});

export default router;