# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: Browse open-source recipes
      parameters:
        - name: query
          in: query
          required: false
          schema: { type: string, maxLength: 120 }
        - name: category
          in: query
          required: false
          schema: { type: string, maxLength: 80 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 30, default: 12 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Recipe cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeList"
  /v1/recipes/{recipeId}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get an open-source recipe detail
      parameters:
        - name: recipeId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Recipe detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RecipeList:
      type: object
      required: [recipes, source]
      properties:
        source: { type: string }
        recipes:
          type: array
          items: { $ref: "#/components/schemas/Recipe" }
        warmupPending:
          type: boolean
          description: >
            True while the server's background nutrition warm-up is still
            running.  Clients should refetch the list shortly so recipe cards
            can show calorie estimates as soon as they are available.
    Recipe:
      type: object
      required: [id, name, source, sourceUrl]
      properties:
        id: { type: string }
        name: { type: string }
        image: { type: string, format: uri, nullable: true }
        category: { type: string, nullable: true }
        area: { type: string, nullable: true }
        description: { type: string, nullable: true }
        instructions: { type: string, nullable: true }
        ingredients:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        prepMinutes: { type: integer, nullable: true }
        calories: { type: number, nullable: true }
        proteinG: { type: number, nullable: true }
        carbsG: { type: number, nullable: true }
        fatG: { type: number, nullable: true }
        nutritionUnavailable:
          type: boolean
          description: >
            True when the server attempted AI nutrition estimation but it
            failed (timeout, nonsensical result, or API error). The client
            should surface a clear "Nutrition unavailable" label and offer a
            retry rather than silently showing blanks.
        source: { type: string }
        sourceUrl: { type: string, format: uri }
...