# lib/api-zod/src/generated/types/recipeList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Recipe } from './recipe';

export interface RecipeList {
  source: string;
  recipes: Recipe[];
  /** True while the server's background nutrition warm-up is still running.  Clients should refetch the list shortly so recipe cards can show calorie estimates as soon as they are available. */
  warmupPending?: boolean;
}